move around document typing files for better organization
