+++ backend/src/document-utils/realtimeOnlineUsers.ts
@@ -1,5 +1,5 @@
 import { OnlineEntity, UpdateType } from "@lib/src/realtimeUserTypes";
-import { UserEntity } from "@lib/src/userTypes";
+import { UserEntity } from "@lib/src/UserEntity";
 
 import FirebaseWrapper from "../firebase-utils/FirebaseWrapper";
