+++ backend/src/security-utils/permissionVerification.ts
@@ -1,4 +1,5 @@
-import { DocumentMetadata, ShareStyle, Document } from "@lib/src/documentTypes";
+import { DocumentMetadata, ShareStyle } from "@lib/src/documentProperties";
+import { Document } from "@lib/src/Document";
 
 // returns true iff a user has read access
 // takes in the email of a user and the document to check
