+++ lib/src/UserEntity.ts
